rename and move file
